test: split error-swallowing suite into its own module
